Added ruby dependency download prerequisite check

+++ joern-cli/frontends/rubysrc2cpg/src/main/scala/io/joern/rubysrc2cpg/RubySrc2Cpg.scala
@@ -12,8 +12,8 @@ import io.joern.x2cpg.utils.ExternalCommand
 import io.shiftleft.codepropertygraph.Cpg
 import io.shiftleft.codepropertygraph.generated.Languages
 import io.shiftleft.passes.CpgPassBase
-import org.jruby.embed.{LocalContextScope, ScriptingContainer}
 import org.slf4j.LoggerFactory
+import sys.process._
 
 import java.nio.file.{Files, Paths}
 import scala.util.{Failure, Success, Try}
@@ -29,26 +29,31 @@ class RubySrc2Cpg extends X2CpgFrontend[Config] {
       val packageTableInfo = new PackageTable()
       new MetaDataPass(cpg, Languages.RUBYSRC, config.inputPath).createAndApply()
       new ConfigFileCreationPass(cpg).createAndApply()
-      //if (config.enableDependencyDownload && !scala.util.Properties.isWin) {
-        val tempDir = File.newTemporaryDirectory()
-        try {
-          downloadDependency(config.inputPath, tempDir.toString())
-          // download(config.inputPath, tempDir.toString())
-          new AstPackagePass(cpg, tempDir.toString(), global, packageTableInfo, config.inputPath).createAndApply()
-        } catch {
-          case ex: Exception =>
-            println(s"Error while parsing dependency: ${ex.getMessage}")
-        } finally {
-          tempDir.delete()
-        }
-      //}
+      downloadDependencies(config, cpg, packageTableInfo)
       val astCreationPass = new AstCreationPass(config.inputPath, cpg, global, packageTableInfo)
       astCreationPass.createAndApply()
       TypeNodePass.withRegisteredTypes(astCreationPass.allUsedTypes(), cpg).createAndApply()
     }
   }
 
-  private def downloadDependency(inputPath: String, tempPath: String): Unit = {
+  private def downloadDependencies(config: Config, cpg: Cpg, packageTable: PackageTable): Unit = {
+    if (config.enableDependencyDownload && !scala.util.Properties.isWin) {
+      if (checkDownloadPrerequisite()) {
+        val tempDir = File.newTemporaryDirectory()
+        try {
+          downloadExternalGemFileDependency(config.inputPath, tempDir.toString())
+          new AstPackagePass(cpg, tempDir.toString(), global, packageTable, config.inputPath).createAndApply()
+        } catch {
+          case ex: Exception =>
+            println(s"Error while parsing dependency: ${ex.getMessage}")
+        } finally {
+          tempDir.delete()
+        }
+      }
+    }
+  }
+
+  private def downloadExternalGemFileDependency(inputPath: String, tempPath: String): Unit = {
     if (Files.isRegularFile(Paths.get(s"${inputPath}${java.io.File.separator}Gemfile"))) {
       ExternalCommand.run(s"bundle config set --local path ${tempPath}", inputPath) match {
         case Success(configOutput) =>
@@ -66,19 +71,21 @@ class RubySrc2Cpg extends X2CpgFrontend[Config] {
     }
   }
 
-//  private def download(inputPath: String, tempPath: String): Unit = {
-//
-//    val container = new ScriptingContainer(LocalContextScope.SINGLETHREAD)
-//    container.getLoadPaths.add(inputPath)
-//    container.runScriptlet(s"require 'rubygems'")
-//    container.runScriptlet(s"require 'bundler'")
-//
-//    val bundlerScript = "require 'bundler'; Bundler.setup"
-//
-//    val gemfileContent = scala.io.Source.fromFile(inputPath).mkString
-//
-//    container.runScriptlet(bundlerScript, gemfileContent)
-//  }
+  private def checkDownloadPrerequisite(): Boolean = {
+    val isRubyInstalled   = "ruby -v".! == 0
+    val isBundleInstalled = "bundle -v".! == 0
+
+    (isRubyInstalled, isBundleInstalled) match {
+      case (false, true) =>
+        logger.error("Skipping Dependency Download: Ruby is not installed.")
+      case (true, false) =>
+        logger.error("Skipping Dependency Download: Bundler is not installed.")
+      case _ =>
+        logger.error("Skipping Dependency Download: Ruby and Bundler are not installed.")
+    }
+
+    isRubyInstalled && isBundleInstalled
+  }
 }
 
 object RubySrc2Cpg {
